pyproject.toml
[tool.poetry]
name = "prose"
version = "3.2.3"
description = "Modular image processing pipelines for Astronomy"
authors = ["Lionel Garcia"]
license = "MIT"
readme = "README.md"
include = ["prose/io/*.sql"]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
numpy = "^1.23.5"
astropy = "^5.1.1"
astroquery = "^0.4.6"
requests = "^2.31.0"
ipython = "*"
scipy = "*"
matplotlib = "*"
scikit-image = "^0.21.0"
pandas = ">1.1"
tqdm = "*"
photutils = "^1.6.0"
pyyaml = "*"
tabulate = "*"
sep = "*"
celerite2 = "*"
twirl = "0.4.0"
multiprocess = "*"
pytest = "*"
imageio = { version = "*", extras = ["ffmpeg"] }

[tool.poetry.group.dev.dependencies]
pytest = "*"
black = "*"

[tool.poetry.group.jax.dependencies]
jax = "*"
jaxopt = "*"

[tool.poetry.group.docs.dependencies]
sphinx = "*"
docutils = "*"
jupyterlab = "*"
myst-parser = "*"
sphinx-book-theme = "^1.0.0"
myst-nb = "*"
sphinx-copybutton = "*"
sphinx-design = "*"
toml = "*"
ipywidgets = "*"
'sphinxcontrib.datatemplates' = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
prose = 'prose.cli.__init__:main'
